+++ assets/javascripts/kitten/index.js
@@ -53,8 +53,6 @@ export { VerticalCard } from './components/cards/vertical-card'
 // Carousel
 export { Carousel } from './components/carousel/carousel'
 export { NavTabCarousel } from './components/carousel/nav-tab-carousel'
-export { CarouselInner } from './components/carousel/carousel-inner'
-export { CarouselPage } from './components/carousel/carousel-page'
 export { SimpleCarousel } from './components/carousel/simple-carousel'
 
 // Comment
@@ -174,6 +172,7 @@ export { Overlay } from './components/modals/overlay'
 export { HorizontalNav } from './components/navigation/horizontal-nav'
 export { Pagination } from './components/navigation/pagination'
 export { TabBar } from './components/navigation/tab-bar'
+export { NavBar } from './components/navigation/nav-bar'
 
 // Notifications
 export { Alert } from './components/notifications/alert'