+++ assets/javascripts/kitten/karl/heroes/hero.js
@@ -4,7 +4,6 @@ import { Title } from '../../components/typography/title'
 import { Paragraph } from '../../components/typography/paragraph'
 import { Marger } from '../../components/layout/marger'
 import { Button } from '../../components/buttons/button'
-import { StyleRoot } from 'radium'
 import { Grid, GridCol } from '../../components/grid/grid'
 
 const lorem = 'Lorem ipsum dolor sit amet.'
@@ -24,68 +23,66 @@ const loremIpsumSitAmet =
 const imageSrc = 'https://placehold.it/40x40/caf4fe/caf4fe'
 
 export const KarlHeroExamples = () => (
-  <StyleRoot>
-    <Grid>
-      <GridCol>
-        <div className="karl-Example">
-          <p className="karl-Example__title">Hero left</p>
-        </div>
-        <Hero imageSrc={imageSrc} direction="left">
-          <Marger bottom="5">
-            <Title modifier="secondary" margin={false}>
-              {lorem}
-            </Title>
-          </Marger>
+  <Grid>
+    <GridCol>
+      <div className="karl-Example">
+        <p className="karl-Example__title">Hero left</p>
+      </div>
+      <Hero imageSrc={imageSrc} direction="left">
+        <Marger bottom="5">
+          <Title modifier="secondary" margin={false}>
+            {lorem}
+          </Title>
+        </Marger>
 
-          <Marger top="5" bottom={4}>
-            <Paragraph modifier="primary" margin={false}>
-              {loremIpsumSitAmet}
-            </Paragraph>
-          </Marger>
+        <Marger top="5" bottom={4}>
+          <Paragraph modifier="primary" margin={false}>
+            {loremIpsumSitAmet}
+          </Paragraph>
+        </Marger>
 
-          <Marger top="4">
-            <Button modifier="helium">Button</Button>
-          </Marger>
-        </Hero>
+        <Marger top="4">
+          <Button modifier="helium">Button</Button>
+        </Marger>
+      </Hero>
 
-        <div className="karl-Example">
-          <p className="karl-Example__title">Hero right</p>
-        </div>
-        <Hero imageSrc={imageSrc} direction="right">
-          <Marger bottom="5">
-            <Title modifier="secondary" margin={false}>
-              {lorem}
-            </Title>
-          </Marger>
+      <div className="karl-Example">
+        <p className="karl-Example__title">Hero right</p>
+      </div>
+      <Hero imageSrc={imageSrc} direction="right">
+        <Marger bottom="5">
+          <Title modifier="secondary" margin={false}>
+            {lorem}
+          </Title>
+        </Marger>
 
-          <Marger top="5" bottom={4}>
-            <Paragraph modifier="primary" margin={false}>
-              {loremIpsumSitAmet}
-            </Paragraph>
-          </Marger>
+        <Marger top="5" bottom={4}>
+          <Paragraph modifier="primary" margin={false}>
+            {loremIpsumSitAmet}
+          </Paragraph>
+        </Marger>
 
-          <Marger top="4">
-            <Button modifier="helium">Button</Button>
-          </Marger>
-        </Hero>
+        <Marger top="4">
+          <Button modifier="helium">Button</Button>
+        </Marger>
+      </Hero>
 
-        <div className="karl-Example">
-          <p className="karl-Example__title">Hero with small text</p>
-        </div>
-        <Hero imageSrc={imageSrc} direction="left" tiny>
-          <Marger bottom="5">
-            <Title modifier="secondary" margin={false}>
-              {lorem}
-            </Title>
-          </Marger>
+      <div className="karl-Example">
+        <p className="karl-Example__title">Hero with small text</p>
+      </div>
+      <Hero imageSrc={imageSrc} direction="left" tiny>
+        <Marger bottom="5">
+          <Title modifier="secondary" margin={false}>
+            {lorem}
+          </Title>
+        </Marger>
 
-          <Marger top="5">
-            <Paragraph modifier="primary" margin={false}>
-              {loremIpsum}
-            </Paragraph>
-          </Marger>
-        </Hero>
-      </GridCol>
-    </Grid>
-  </StyleRoot>
+        <Marger top="5">
+          <Paragraph modifier="primary" margin={false}>
+            {loremIpsum}
+          </Paragraph>
+        </Marger>
+      </Hero>
+    </GridCol>
+  </Grid>
 )