+++ src/components/cards/reward-card/components/checked-section.js
@@ -1,0 +1,29 @@
+"use strict";
+
+var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefault");
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.RewardCardCheckedSection = void 0;
+
+var _extends2 = _interopRequireDefault(require("@babel/runtime/helpers/extends"));
+
+var _objectWithoutProperties2 = _interopRequireDefault(require("@babel/runtime/helpers/objectWithoutProperties"));
+
+var _react = _interopRequireDefault(require("react"));
+
+var _classnames = _interopRequireDefault(require("classnames"));
+
+var _checkedIconLine = require("./checked-icon-line");
+
+var RewardCardCheckedSection = function RewardCardCheckedSection(_ref) {
+  var children = _ref.children,
+      className = _ref.className,
+      props = (0, _objectWithoutProperties2.default)(_ref, ["children", "className"]);
+  return /*#__PURE__*/_react.default.createElement(_react.default.Fragment, null, /*#__PURE__*/_react.default.createElement(_checkedIconLine.RewardCardCheckedIconLine, null), /*#__PURE__*/_react.default.createElement("div", (0, _extends2.default)({
+    className: (0, _classnames.default)('k-RewardCard__checkedSection', className)
+  }, props), children));
+};
+
+exports.RewardCardCheckedSection = RewardCardCheckedSection;